Workaround super awkward pre-memory-color.
Not sure if this is a hardware bug. Seems very specific that memory
color is only coherent in cycle1, but memory coverage is read early, so
it seems rather strange that things are pipelined like this.
Either way, fixing this is purely to workaround games which (by mistake) use this behavior.